Add necessary files, working

+++ baselines/flwr_baselines/publications/fl_wav2vec2/data_prepare.py
@@ -55,31 +55,45 @@ def delete_file(filename):
     print(f"Deleted {filename}.")
 
 #Change the path corespond to your actual path    
-def csv_path_audio(path):
+def csv_path_audio():
+    df_concat_train = None
+    df_concat_dev = None
+    df_concat_test = None
     for i in range(1983):
-        df_train = pd.read_csv(f"./data/client_{i}/ted_train.csv")
-        df_dev = pd.read_csv(f"./data/client_{i}/ted_dev.csv")
-        df_test = pd.read_csv(f"./data/client_{i}/ted_test.csv")
-        df_train["wav"] = df_train["wav"].str.replace("/local_disk/idyie/tnguyen/TEDLIUM_release-3/legacy/train/sph/",path)
-        df_dev["wav"] = df_dev["wav"].str.replace("/local_disk/idyie/tnguyen/TEDLIUM_release-3/legacy/train/sph/",path)
-        df_test["wav"] = df_test["wav"].str.replace("/local_disk/idyie/tnguyen/TEDLIUM_release-3/legacy/train/sph/",path)
-        df_train.to_csv(f"./data/client_{i}/ted_train.csv",index = False)
-        df_dev.to_csv(f"./data/client_{i}/ted_dev.csv", index = False)
-        df_test.to_csv(f"./data/client_{i}/ted_test.csv", index = False)
+        df_train = pd.read_csv(f"./ressources/client_{i}/ted_train.csv")
+        df_dev = pd.read_csv(f"./ressources/client_{i}/ted_dev.csv")
+        df_test = pd.read_csv(f"./ressources/client_{i}/ted_test.csv")
+        if df_concat_train is None:
+            df_concat_train = df_train
+        else:
+            df_concat_train = pd.concat([df_concat_train, df_train], ignore_index=True)
+        if df_concat_dev is None:
+            df_concat_dev = df_dev
+        else:
+            df_concat_dev = pd.concat([df_concat_dev, df_dev], ignore_index=True)
+        if df_concat_test is None:
+            df_concat_test = df_test
+        else:
+            df_concat_test = pd.concat([df_concat_test, df_test], ignore_index=True)
+
+    df_concat_train.to_csv(f"./ressources/ted_train.csv",index = False)
+    df_concat_dev.to_csv(f"./ressources/ted_dev.csv", index = False)
+    df_concat_test.to_csv(f"./ressources/ted_test.csv", index = False)
+
 
 # CHANGE THE PATH CORESPOND TO YOUR PATH
 url = "https://projets-lium.univ-lemans.fr/wp-content/uploads/corpus/TED-LIUM/TEDLIUM_release-3.tgz"
-filename = "data/TEDLIUM_release-3.tgz"
-extract_path = "data/audio" # replace with the path to the directory where you want to extract the files
+filename = "ressources/TEDLIUM_release-3.tgz"
+extract_path = "ressources/audio" # replace with the path to the directory where you want to extract the files
 
 
-try:
-    download_file(url, filename)
-    extract_file(filename, extract_path)
-finally:
-    delete_file(filename)
+if not os.path.exists(extract_path):
+    try:
+        download_file(url, filename)
+        extract_file(filename, extract_path)
+    finally:
+        delete_file(filename)
 
-
-csv_path_audio(extract_path)
+csv_path_audio()
